contentCheck/requirements.txt
opencv-python>=4.8.0
mediapipe>=0.10.9
ultralytics>=8.3.0
google-genai>=1.0.0
openai>=1.30.0
numpy>=1.24.0
Pillow>=10.0.0
